Added a server file
