+++ Gemfile
@@ -3,3 +3,5 @@ source "https://rubygems.org"
 
 gem "dotenv"
 gem "httparty"
+gem "sinatra"
+gem 'sinatra-cross_origin'
