+++ README.md
@@ -1,1 +1,6 @@
 # project3
+
+# Back-End Setup
+Install sinatra, dotenv, httparty, and sinatra-cross-origin
+
+Then run ruby server.rb to start the server
